fix: 잘못 파싱하던 자료형 수정

+++ lib/features/workout_history/domain/entities/workout_activity.dart
@@ -18,7 +18,7 @@ class WorkoutActivity {
   final DateTime endedAt;
   final double distance;
   final int duration; // seconds
-  final int elevationGain; // meters
+  final double elevationGain; // meters
   final String thumbnailImageUrl;
   final String userProfileImageUrl;
   final String userNickname;
